requirements.txt
colorama
pypdf2
docx
python-docx
openai
dotenv
passlib[bcrypt]
python-jose
fastapi
psycopg2-binary
uvicorn
redis
sqlalchemy
pydantic
pydantic[email]
pydantic-settings
pymongo
python-multipart
numpy
pytest
pytest-asyncio
httpx
mongomock
sqlalchemy